move tracer to own file
